backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
sentence-transformers==2.2.2
hdbscan==0.8.29
scikit-learn==1.3.0
numpy==1.24.3
konlpy==0.6.0
keybert==0.7.0
huggingface-hub==0.16.4
transformers==4.21.3
torch==1.13.1
gensim==4.3.2